requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==6.0.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.1.3
packaging==24.2
peewee==3.17.9
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.17.0
tomli==2.2.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0

